Moves files into server directory
